+++ INUBus/Engineering/Test3.swift
@@ -1,0 +1,19 @@
+//
+//  Test3.swift
+//  INUBus
+//
+//  Created by zun on 30/08/2019.
+//  Copyright © 2019 zun. All rights reserved.
+//
+
+import UIKit
+
+class Test3: Test1 {
+  override var busStopIdentifier: String {
+    return "science"
+  }
+  
+  override func viewDidLoad() {
+    super.viewDidLoad()
+  }
+}
