+++ INUBus/ViewController.swift
@@ -8,10 +8,10 @@
 
 import UIKit
 
-class ViewController: UIViewController {
+class ViewController: UITabBarController {
   
   override func viewDidLoad() {
-    super.viewDidLoad()
+    
   }
 
 }
